moved setup script to library to share it between notebooks and library
